Update styling format of the card
By adding a width and height to the button we get a round circle when providing border-radius of 100%. This trumps the oval shape.
Also, giving the icon inside the button a position property and giving it a top and left attribute will create the centered effect we are aiming for. I hope this code is helpful!